+++ miataru/miataru/views/iPhone/iPhone_RootView.swift
@@ -108,6 +108,9 @@ struct iPhone_RootView: View {
 
     private func applyRootNavigationDestination(_ destination: AppRootNavigationDestination?) {
         switch destination {
+        case .devices:
+            selectedTab = 0
+            appNavigation.consumeRootDestination(.devices)
         case .settings:
             selectedTab = 2
         case .none: